fix sequelize
