+++ .env.example
@@ -9,3 +9,6 @@ DB_PROD_USERNAME="USER NAME"
 DB_PROD_PASSWORD="PASSW0RD"
 DB_PROD_NAME=DB_PROD_NAME
 DB_PROD_HOST="DB_PROD_HOST"
+
+
+
