+++ src/database/models/contact.js
@@ -1,0 +1,24 @@
+'use strict';
+const {
+  Model
+} = require('sequelize');
+module.exports = (sequelize, DataTypes) => {
+  class Contact extends Model {
+    /**
+     * Helper method for defining associations.
+     * This method is not a part of Sequelize lifecycle.
+     * The `models/index` file will call this method automatically.
+     */
+    static associate(models) {
+      // define association here
+    }
+  };
+  Contact.init({
+    name: DataTypes.STRING,
+    number: DataTypes.INTEGER
+  }, {
+    sequelize,
+    modelName: 'Contact',
+  });
+  return Contact;
+};
