--- src/models/contact.js
@@ -1,15 +1,0 @@
-"use strict";
-module.exports = (sequelize, DataTypes) => {
-  const Contact = sequelize.define(
-    "Contact",
-    {
-      name: DataTypes.STRING,
-      number: DataTypes.STRING,
-    },
-    {}
-  );
-  Contact.associate = function (models) {
-    // associations can be defined here
-  };
-  return Contact;
-};
